pyproject.toml
[tool.poetry]
name = "volby-2023"
version = "0.1.0"
description = ""
authors = ["Roman Zydyk <romanzdk@gmail.com>"]
readme = "README.md"
packages = [{include = "volby_2023"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.4"
python-decouple = "^3.6"
streamlit = "^1.7.0"
psycopg2 = "^2.9.5"
plotly = "^5.11.0"
streamlit-autorefresh = "^0.0.1"
geopandas = "^0.12.2"
matplotlib = "^3.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
